auth Pog

+++ resources/post.py
@@ -1,3 +1,4 @@
+from middleware import read_token, strip_token
 from models.db import db
 from models.post import *
 from flask_restful import Resource
@@ -7,24 +8,36 @@ from sqlalchemy.orm import joinedload
 
 class Posts(Resource):
     def get(self):
-        posts = Post.find_all()
-        return posts
+        token = strip_token(request)
+        if read_token(token)['data']:
+            posts = Post.find_all()
+            return posts
+        else:
+            return read_token(token)['payload']
 
     def post(self):
-        data = request.get_json()
-        params = {}
-        for k in data.keys():
-            params[k] = data[k]
-        post = Post(**params)
-        post.create()
-        return post.json(), 201
+        token = strip_token(request)
+        if read_token(token)['data']:
+            data = request.get_json()
+            params = {}
+            for k in data.keys():
+                params[k] = data[k]
+            post = Post(**params)
+            post.create()
+            return post.json(), 201
+        else:
+            return read_token(token)['payload']
 
 
 class PostDetails(Resource):
     def get(self, post_id):
-        post = Post.query.options(joinedload(
-            "user")).filter_by(id=post_id).first()
-        return {**post.json(), "user": post.user.json()}
+        token = strip_token(request)
+        if read_token(token)['data']:
+            post = Post.query.options(joinedload(
+                "user")).filter_by(id=post_id).first()
+            return {**post.json(), "user": post.user.json()}
+        else:
+            return read_token(token)['payload']
 
     def put(self, post_id):
         data = request.get_json()
@@ -35,10 +48,13 @@ class PostDetails(Resource):
         return post.json()
 
     def delete(self, post_id):
-        print(post_id)
-        post = Post.find_by_id(post_id)
-        if not post:
-            return {"Message": "Not Found"}, 404
-        db.session.delete(post)
-        db.session.commit()
-        return {"Message": "Post Deleted", "payload": post_id}
+        token = strip_token(request)
+        if read_token(token)['data']:
+            post = Post.find_by_id(post_id)
+            if not post:
+                return {"Message": "Not Found"}, 404
+            db.session.delete(post)
+            db.session.commit()
+            return {"Message": "Post Deleted", "payload": post_id}
+        else:
+            return read_token(token)['payload']
